fix installation: move main files to another directory
