+++ cpp_stats/main.py
@@ -3,7 +3,7 @@ Main file of project.
 '''
 
 import click
-from cpp_stats import cpp_stats
+from cpp_stats.cpp_stats import CppStats
 
 @click.command()
 @click.option('--report', default='report.xml',
@@ -18,7 +18,7 @@ def main(path_to_repo, report):
     path_to_repo (str): Path to the repository.
     report (str): Path to xml file where to store report.
     '''
-    stats = cpp_stats.CppStats(path_to_repo, True)
+    stats = CppStats(path_to_repo, True)
     with open(report, "w", encoding="utf-8") as f:
         click.echo(stats.as_xml(), file=f)
