+++ cpp_stats/metrics/cyclomatic_complexity.py
