+++ cpp_stats/metrics/halstead/delivered_bugs.py
