+++ cpp_stats/metrics/halstead/difficulty.py
