+++ cpp_stats/metrics/lcom/base.py
