+++ cpp_stats/metrics/lcom/base_metric.py
