+++ cpp_stats/metrics/lcom/lcom1.py
