+++ cpp_stats/metrics/lcom/lcom2.py
