+++ cpp_stats/metrics/lcom/lcom3.py
