+++ cpp_stats/metrics/number_of_classes.py
