+++ cpp_stats/metrics/number_of_methods.py
