+++ pytest.ini
@@ -1,2 +1,2 @@
-[pytest]
-pythonpath = src
+; [pytest]
+; pythonpath = ./cpp_stats
